requirements.txt
streamlit>=1.31.0
Pillow>=10.1.0
numpy>=1.26.0